pyproject.toml
[project]
name = "fap"
version = "0.1.0"
description = "API for AST"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "fastapi[standard]>=0.116.1",
    "h2>=4.3.0",
    "openai>=2.15.0",
    "pillow>=10.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.0.0",
    "speechmatics-rt>=0.5.3",
    "sqlalchemy>=2.0.43",
    "websockets>=12.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
